// .repos/semantic-web/traqula/packages/core/lib/types.ts
export interface Typed<Type extends string = string> {
  type: Type;
  subType?: string;
}

/**
 * A {@link Typed} object that is sure to have a subType
 */
export interface SubTyped<Type extends string = string, SubType extends string = string> extends Typed<Type> {
  subType: SubType;
}